[MBUILDCACHE-99] Project checksum fails to identify moved files in certain cases (#161)
* [MBUILDCACHE-99] Project checksum fails to identify moved files in certain cases.

Input files checksum is calculated by sorting all files and using their content (but not their path) to update a checksum. This means that if a file is moved / renamed in a way that doesn't change its location in the list of sorted files, the input files checksum will not change.
This change adds the file's path to the checksum.